MYR-729: moved styles to separate file
