+++ src/components-v2/atoms/avatar/custom-avatar.style.ts
@@ -1,0 +1,20 @@
+import {createStyles, makeStyles, Theme} from '@material-ui/core/styles';
+
+const useStyles = makeStyles((theme: Theme) =>
+  createStyles({
+    small: {
+      width: theme.spacing(3.75),
+      height: theme.spacing(3.75),
+    },
+    medium: {
+      width: theme.spacing(6),
+      height: theme.spacing(6),
+    },
+    large: {
+      width: theme.spacing(9),
+      height: theme.spacing(9),
+    },
+  }),
+);
+
+export default useStyles;
