+++ src/components-v2/atoms/avatar/index.ts
@@ -1,5 +1,6 @@
 import {default as CustomAvatar} from './CustomAvatar';
 import {CustomAvatarSize, CustomAvatarProps} from './custom-avatar.interfaces';
+import useStyles from './custom-avatar.style';
 
-export {CustomAvatar, CustomAvatarSize};
+export {CustomAvatar, CustomAvatarSize, useStyles};
 export type {CustomAvatarProps};
